Switch to a flat layout for Elements
This commit switched the cache to using a flat layout for Elements;
instead of storing the Elements in a linked list, it stores them in a
slice, and stores indexes into the slice where it once stored pointers.
This has two advantages:

  1. The resulting code is slightly smaller.
  2. It has better locality

One issue the linked list version could have is that after extensive
deletions and additions, the Elements could be scattered across memory.
By storing the elements in slices, and reusing the slots when elements
are removed, all the elements should remain close together.

This strategy has the disadvantage in that it might make eviction
slower; it depends on if the better locality or the number of elements
to check is the dominant term.